pyproject.toml
[project]
name = "metapromptingframework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0",
]
